rebuild-kernel: use native GOARCH for Docker
When building a kernel inside a containerized Debian, use the native
GOARCH version of Debian instead of arm64. This allows to use

Note that this change requires a gokr-rebuild-kernel binary compiled
for GOOS=linux named gokr-rebuild-kernel.$GOARCH, so changes in
gokrazy/kernel Makefile is required:
  GOOS=linux go build -o gokr-rebuild-kernel.$(go env GOARCH) \
    github.com/gokrazy/autoupdate/cmd/gokr-rebuild-kernel

This change is a step to improve compilation speed on non-linux/amd64
platforms such as macOS on Apple Silicon. This will also open the way
for building a kernel from Windows.